Rearrangements to simplify dependences of aux and messages
