+++ src/lime_config.h
@@ -12,21 +12,12 @@
 
 #include "dims.h"
 
-#define SQRT_PI                 (sqrt(M_PI))           /* sqrt(pi)	*/
-#define ARCSEC_TO_RAD           (M_PI/180.0/3600.0)
-#define EPS                     1.0e-30                /* general use small number */
-
 #define MAX_NSPECIES            100
 #define MAX_NIMAGES             100
 #define NUM_GRID_STAGES         5
 #define MAX_N_COLL_PART         20
-#define TYPICAL_ISM_DENS        1000.0
-#define STR_LEN_0               80
 #define DENSITY_POWER           0.2
 #define MAX_N_HIGH              10
-
-
-extern int silent;
 
 typedef struct {
   /* Elements also present in struct inpars: */
