+++ src/run_new.c
@@ -880,7 +880,7 @@ exit(1);
   /*Set locale to avoid trouble when reading files*/
   setlocale(LC_ALL, "C");
 
-  if(!silent) greetings();
+  if(!silent) greetings(VERSION);
   if(!silent) screenInfo();
 
 #ifdef FASTEXP
